requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.1
Django==5.2.9
django-allauth==65.13.1
django-redis==6.0.0
djangorestframework==3.16.1
idna==3.11
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.2
cryptography
django-storages
boto3
django-ckeditor
ckeditor
ckeditor-
dj-rest-auth
django-allauth
djangorestframework-simplejwt
google-auth

djangorestframework-simplejwt[token_blacklist]
